Move upper hessenberg decomposition into module
